+++ test/schema.test.js
@@ -1981,22 +1981,18 @@ describe('schema', function() {
         assert.ok(clone.path('events').Constructor.discriminators['gh7954_Clicked']);
         assert.ok(clone.path('events').Constructor.discriminators['gh7954_Purchased']);
       });
-    });
 
-    it('TTL index with timestamps (gh-5656)', function(done) {
-      const testSchema = new mongoose.Schema({
-        foo: String,
-        updatedAt: {
-          type: Date,
-          expires: '2h'
-        }
-      }, { timestamps: true });
+      it('uses Mongoose instance\'s Schema constructor (gh-9426)', function() {
+        const db = new mongoose.Mongoose();
+        db.Schema.prototype.localTest = function() {
+          return 42;
+        };
+        const test = new db.Schema({});
+        assert.equal(test.localTest(), 42);
 
-      const indexes = testSchema.indexes();
-      assert.deepEqual(indexes, [
-        [{ updatedAt: 1 }, { background: true, expireAfterSeconds: 7200 }]
-      ]);
-      done();
+        const test2 = test.clone();
+        assert.equal(test2.localTest(), 42);
+      });
     });
 
     it('childSchemas prop (gh-5695)', function(done) {
@@ -2500,5 +2496,14 @@ describe('schema', function() {
     assert.equal(schematype.options.type, Number);
     assert.equal(schematype.options.min, 4);
     assert.equal(schematype.options.get, get);
+  });
+
+  it('applies correct schema to nested primitive arrays (gh-9429)', function() {
+    const schema = new Schema({
+      ids: [[{ type: 'ObjectId', required: true }]]
+    });
+
+    const casted = schema.path('ids').cast([[]]);
+    assert.equal(casted[0].$path(), 'ids.$');
   });
 });